Extract rounding-mode vocabulary into Rounding.mts
The RoundingMode type and mode constants are shared vocabulary, spoken
by both Decimal.mts and CoefficientExponent.mts; parking them in
CoefficientExponent.mts made Decimal.mts depend on the class module for
names the class does not own. Move them (and the ROUNDING_MODES
validation list) into a new internal src/Rounding.mts so each module
declares that dependency explicitly. Like CoefficientExponent.mts, the
new module sits behind the package's exports fence.
