+++ src/CoefficientExponent.mts
@@ -8,18 +8,14 @@
  * @author Jesse Alama <jesse@igalia.com>
  */
 
-export const ROUNDING_MODE_CEILING = "ceil";
-export const ROUNDING_MODE_FLOOR = "floor";
-export const ROUNDING_MODE_TRUNCATE = "trunc";
-export const ROUNDING_MODE_HALF_EVEN = "halfEven";
-export const ROUNDING_MODE_HALF_EXPAND = "halfExpand";
-
-export type RoundingMode =
-    | "ceil"
-    | "floor"
-    | "trunc"
-    | "halfEven"
-    | "halfExpand";
+import {
+    ROUNDING_MODE_CEILING,
+    ROUNDING_MODE_FLOOR,
+    ROUNDING_MODE_HALF_EVEN,
+    ROUNDING_MODE_HALF_EXPAND,
+    ROUNDING_MODE_TRUNCATE,
+    type RoundingMode,
+} from "./Rounding.mjs";
 
 /**
  * Apply a rounding mode to a positive CoefficientExponent value
