+++ stryker.config.mjs
@@ -3,7 +3,7 @@ export default {
     testRunner: "command",
     commandRunner: {
         command:
-            "./node_modules/.bin/esbuild src/Decimal.mts src/CoefficientExponent.mts" +
+            "./node_modules/.bin/esbuild src/Decimal.mts src/CoefficientExponent.mts src/Rounding.mts" +
             " --format=esm --target=es2020 --outdir=src --out-extension:.js=.mjs" +
             " && ./node_modules/.bin/jest",
     },
