+++ tests/encapsulation.test.js
@@ -3,9 +3,12 @@ describe("package encapsulation", () => {
         const mod = await import("proposal-decimal");
         expect(mod.Decimal).toBeDefined();
     });
-    test("internal modules cannot be imported from the package", async () => {
-        await expect(
-            import("proposal-decimal/src/CoefficientExponent.mjs")
-        ).rejects.toThrow(/not defined by "exports"/);
-    });
+    test.each(["CoefficientExponent", "Rounding"])(
+        "internal module %s cannot be imported from the package",
+        async (mod) => {
+            await expect(
+                import(`proposal-decimal/src/${mod}.mjs`)
+            ).rejects.toThrow(/not defined by "exports"/);
+        }
+    );
 });
